+++ public/index.html
@@ -13,7 +13,7 @@
   <body>
     <div id="content"></div>
     <script type="text/babel" src="scripts/example.js"></script>
-    <script type="text/babel" src="scripts/comments.js"></script>
+    
     
     <script type="text/babel"> 
     </script>
