+++ public/scripts/components/commentList.js
@@ -1,0 +1,19 @@
+import Comment from "comment";
+
+
+var CommentList = React.createClass({
+    render: function() {
+      var commentNodes = this.props.data.map(function(comment) {
+        return (
+          <Comment author={comment.author} key={comment.id}>
+            {comment.text}{comment.date}
+          </Comment>
+        );
+      });
+      return (
+        <div className="commentList">
+          {commentNodes}
+        </div>
+      );
+    }
+  });
