+++ server.js
@@ -12,7 +12,7 @@ app.use('/', express.static(path.join(__dirname, 'public')));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({extended: true}));
 
-/
+
 app.use(function(req, res, next) {
     
     res.setHeader('Access-Control-Allow-Origin', '*');
